move funcs to related files
