+++ pkg/minikube/bootstrapper/bsutil/kubelet_test.go
@@ -14,12 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// bsutil package will eventually be renamed to kubeadm package after getting rid of older one
 package bsutil
 
 import (
-	"fmt"
-	"io/ioutil"
-	"strings"
 	"testing"
 
 	"github.com/pmezard/go-difflib/difflib"
@@ -166,200 +164,5 @@ ExecStart=/var/lib/minikube/binaries/v1.17.0/kubelet --authorization-mode=Webhoo
 				t.Errorf("unexpected diff:\n%s", diff)
 			}
 		})
-	}
-}
-
-func getExtraOpts() []config.ExtraOption {
-	return config.ExtraOptionSlice{
-		config.ExtraOption{
-			Component: Apiserver,
-			Key:       "fail-no-swap",
-			Value:     "true",
-		},
-		config.ExtraOption{
-			Component: ControllerManager,
-			Key:       "kube-api-burst",
-			Value:     "32",
-		},
-		config.ExtraOption{
-			Component: Scheduler,
-			Key:       "scheduler-name",
-			Value:     "mini-scheduler",
-		},
-		config.ExtraOption{
-			Component: Kubeadm,
-			Key:       "ignore-preflight-errors",
-			Value:     "true",
-		},
-		config.ExtraOption{
-			Component: Kubeadm,
-			Key:       "dry-run",
-			Value:     "true",
-		},
-	}
-}
-
-func getExtraOptsPodCidr() []config.ExtraOption {
-	return config.ExtraOptionSlice{
-		config.ExtraOption{
-			Component: Kubeadm,
-			Key:       "pod-network-cidr",
-			Value:     "192.168.32.0/20",
-		},
-	}
-}
-
-func recentReleases() ([]string, error) {
-	// test the 6 most recent releases
-	versions := []string{"v1.17", "v1.16", "v1.15", "v1.14", "v1.13", "v1.12", "v1.11"}
-	foundNewest := false
-	foundDefault := false
-
-	for _, v := range versions {
-		if strings.HasPrefix(constants.NewestKubernetesVersion, v) {
-			foundNewest = true
-		}
-		if strings.HasPrefix(constants.DefaultKubernetesVersion, v) {
-			foundDefault = true
-		}
-	}
-
-	if !foundNewest {
-		return nil, fmt.Errorf("No tests exist yet for newest minor version: %s", constants.NewestKubernetesVersion)
-	}
-
-	if !foundDefault {
-		return nil, fmt.Errorf("No tests exist yet for default minor version: %s", constants.DefaultKubernetesVersion)
-	}
-
-	return versions, nil
-}
-
-/**
-Need a separate test function to test the DNS server IP
-as v1.11 yaml file is very different compared to v1.12+.
-This test case has only 1 thing to test and that is the
-nnetworking/dnsDomain value
-*/
-func TestGenerateKubeadmYAMLDNS(t *testing.T) {
-	versions := []string{"v1.16", "v1.15", "v1.14", "v1.13", "v1.12"}
-	tests := []struct {
-		name      string
-		runtime   string
-		shouldErr bool
-		cfg       config.KubernetesConfig
-	}{
-		{"dns", "docker", false, config.KubernetesConfig{DNSDomain: "1.1.1.1"}},
-	}
-	for _, version := range versions {
-		for _, tc := range tests {
-			runtime, err := cruntime.New(cruntime.Config{Type: tc.runtime})
-			if err != nil {
-				t.Fatalf("runtime: %v", err)
-			}
-			tname := tc.name + "_" + version
-			t.Run(tname, func(t *testing.T) {
-				cfg := tc.cfg
-				cfg.NodeIP = "1.1.1.1"
-				cfg.NodeName = "mk"
-				cfg.KubernetesVersion = version + ".0"
-
-				got, err := GenerateKubeadmYAML(cfg, runtime)
-				if err != nil && !tc.shouldErr {
-					t.Fatalf("got unexpected error generating config: %v", err)
-				}
-				if err == nil && tc.shouldErr {
-					t.Fatalf("expected error but got none, config: %s", got)
-				}
-				if tc.shouldErr {
-					return
-				}
-				expected, err := ioutil.ReadFile(fmt.Sprintf("testdata/%s/%s.yaml", version, tc.name))
-				if err != nil {
-					t.Fatalf("unable to read testdata: %v", err)
-				}
-				diff, err := difflib.GetUnifiedDiffString(difflib.UnifiedDiff{
-					A:        difflib.SplitLines(string(expected)),
-					B:        difflib.SplitLines(string(got)),
-					FromFile: "Expected",
-					ToFile:   "Got",
-					Context:  1,
-				})
-				if err != nil {
-					t.Fatalf("diff error: %v", err)
-				}
-				if diff != "" {
-					t.Errorf("unexpected diff:\n%s\n===== [RAW OUTPUT] =====\n%s", diff, got)
-				}
-			})
-		}
-	}
-}
-
-func TestGenerateKubeadmYAML(t *testing.T) {
-	extraOpts := getExtraOpts()
-	extraOptsPodCidr := getExtraOptsPodCidr()
-	versions, err := recentReleases()
-	if err != nil {
-		t.Errorf("versions: %v", err)
-	}
-	tests := []struct {
-		name      string
-		runtime   string
-		shouldErr bool
-		cfg       config.KubernetesConfig
-	}{
-		{"default", "docker", false, config.KubernetesConfig{}},
-		{"containerd", "containerd", false, config.KubernetesConfig{}},
-		{"crio", "crio", false, config.KubernetesConfig{}},
-		{"options", "docker", false, config.KubernetesConfig{ExtraOptions: extraOpts}},
-		{"crio-options-gates", "crio", false, config.KubernetesConfig{ExtraOptions: extraOpts, FeatureGates: "a=b"}},
-		{"unknown-component", "docker", true, config.KubernetesConfig{ExtraOptions: config.ExtraOptionSlice{config.ExtraOption{Component: "not-a-real-component", Key: "killswitch", Value: "true"}}}},
-		{"containerd-api-port", "containerd", false, config.KubernetesConfig{NodePort: 12345}},
-		{"containerd-pod-network-cidr", "containerd", false, config.KubernetesConfig{ExtraOptions: extraOptsPodCidr}},
-		{"image-repository", "docker", false, config.KubernetesConfig{ImageRepository: "test/repo"}},
-	}
-	for _, version := range versions {
-		for _, tc := range tests {
-			runtime, err := cruntime.New(cruntime.Config{Type: tc.runtime})
-			if err != nil {
-				t.Fatalf("runtime: %v", err)
-			}
-			tname := tc.name + "_" + version
-			t.Run(tname, func(t *testing.T) {
-				cfg := tc.cfg
-				cfg.NodeIP = "1.1.1.1"
-				cfg.NodeName = "mk"
-				cfg.KubernetesVersion = version + ".0"
-
-				got, err := GenerateKubeadmYAML(cfg, runtime)
-				if err != nil && !tc.shouldErr {
-					t.Fatalf("got unexpected error generating config: %v", err)
-				}
-				if err == nil && tc.shouldErr {
-					t.Fatalf("expected error but got none, config: %s", got)
-				}
-				if tc.shouldErr {
-					return
-				}
-				expected, err := ioutil.ReadFile(fmt.Sprintf("testdata/%s/%s.yaml", version, tc.name))
-				if err != nil {
-					t.Fatalf("unable to read testdata: %v", err)
-				}
-				diff, err := difflib.GetUnifiedDiffString(difflib.UnifiedDiff{
-					A:        difflib.SplitLines(string(expected)),
-					B:        difflib.SplitLines(string(got)),
-					FromFile: "Expected",
-					ToFile:   "Got",
-					Context:  1,
-				})
-				if err != nil {
-					t.Fatalf("diff error: %v", err)
-				}
-				if diff != "" {
-					t.Errorf("unexpected diff:\n%s\n===== [RAW OUTPUT] =====\n%s", diff, got)
-				}
-			})
-		}
 	}
 }
